LLM-and-Recommendation-Sys/rag/requirements.txt
langchain-core>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.15
langchain-openai>=0.0.3
openai>=1.5.0
psycopg2-binary>=2.9.9
pgvector>=0.2.3
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
